Move some definitions to the new `math` module
