+++ src/hoa/mod.rs
@@ -12,7 +12,7 @@ use hoars::{HoaAutomaton, ALPHABET, MAX_APS, VARS};
 /// Assume we have a propositional alphabet over the atomic propositions `a`, `b` and `c`.
 ///
 /// Then a **symbol** in this alphabet is a valuation of these variables, e.g. `a & !b & c`. This is used to label
-/// transitions in a [`crate::ts::TransitionSystem`].
+/// transitions in a [`crate::transition_system::TransitionSystem`].
 ///
 /// An **expression** on the other hand is used to label edges and it is a boolean expression over
 /// the atomic propositions, e.g. `(a | b) & c`. Such an expression is matched by
