+++ src/transition_system/operations/restricted.rs
@@ -1,6 +1,6 @@
 use std::marker::PhantomData;
 
-use crate::{prelude::*, Set};
+use crate::{math::Set, prelude::*};
 
 /// Abstracts the filtering of a transition system's state indices. This trait is implemented by
 /// functions which take a state index and return a boolean value indicating whether the state index
