+++ src/transition_system/sproutable.rs
@@ -1,7 +1,7 @@
 use bit_set::BitSet;
 use itertools::Itertools;
 
-use crate::{prelude::*, Bijection};
+use crate::{math::Bijection, prelude::*};
 
 /// Marker trait for [`Alphabet`]s that can be indexed, i.e. where we can associate each
 /// [`Alphabet::Symbol`] and [`Alphabet::Expression`] with a unique index (a `usize`).
@@ -66,7 +66,7 @@ pub trait Sproutable: TransitionSystem {
     /// ```
     /// use automata::prelude::*;
     ///
-    /// let source = TSBuilder::default()
+    /// let source: DTS<CharAlphabet, usize, usize> = TSBuilder::default()
     ///     .with_transitions([(0, 'a', 0, 0), (0, 'b', 0, 0)])
     ///     .with_colors([0])
     ///     .deterministic();
